+++ video/src/main/scala/net/hearthstats/video/SequenceEncoder.scala
@@ -1,128 +1,95 @@
-package net.hearthstats.video
+package net.hearthstats.modules.video
 
 import java.awt.geom.AffineTransform
 import java.awt.image.{ AffineTransformOp, BufferedImage }
-import scala.concurrent.duration.DurationInt
 import java.io.File
-import java.nio.ByteBuffer
-import java.util.{ ArrayList, Arrays }
 import scala.concurrent.{ Future, Promise }
 import scala.concurrent.ExecutionContext.Implicits.global
-import org.jcodec.codecs.h264.{ H264Encoder, H264Utils }
-import org.jcodec.common.NIOUtils
-import org.jcodec.common.model.{ ColorSpace, Picture }
-import org.jcodec.containers.mp4.{ Brand, MP4Packet, TrackType }
-import org.jcodec.containers.mp4.muxer.MP4Muxer
-import org.jcodec.scale.{ AWTUtil, RgbToYuv420 }
-import com.xuggle.mediatool.ToolFactory
+import com.xuggle.mediatool.{ IMediaWriter, ToolFactory }
 import grizzled.slf4j.Logging
+import java.util.concurrent.TimeUnit
+import javax.imageio.ImageIO
+import com.xuggle.xuggler.IRational
+import scala.util.control.NonFatal
 import akka.actor.ActorSystem
-import akka.actor.Props
 import akka.actor.Actor
+import akka.actor.Props
 import akka.pattern.ask
 import akka.util.Timeout
-import java.nio.file.Files
-import javax.imageio.ImageIO
+import scala.concurrent.duration._
+import akka.pattern.ask
+import com.xuggle.xuggler.video.ConverterFactory
+import com.xuggle.xuggler.IPixelFormat
+import com.xuggle.xuggler.video.ArgbConverter
 
-class SequenceEncoder(
-  framesPerSec: Integer = 10,
-  videoWidth: Integer = 800,
-  videoHeight: Integer = 600) extends VideoEncoder with Logging {
+class SequenceEncoder extends VideoEncoder with Logging {
+  val system = ActorSystem("video")
 
-  val out: File = File.createTempFile("HSReplay", "video.mp4")
-  val picFolder = Files.createTempDirectory("hspic")
-  info(s"Storing pictures in $picFolder")
+  def newVideo(framesPerSec: Double, videoWidth: Int, videoHeight: Int) = new OngoingVideo {
+    val actor = system.actorOf(Props(new EncodeActor))
 
-  val ch = NIOUtils.writableFileChannel(out)
-  val transform = new RgbToYuv420(0, 0)
-  val encoder = new H264Encoder
-  val spsList = new ArrayList[ByteBuffer]()
-  val ppsList = new ArrayList[ByteBuffer]()
-  val _out = ByteBuffer.allocate(1920 * 1080 * 6)
-  var frameNo = 0
-  var receivedFrames = 0
-  var storedFrames = 0
-  val muxer = new MP4Muxer(ch, Brand.MP4)
-  val outTrack = muxer.addTrackForCompressed(TrackType.VIDEO, framesPerSec)
-  var closed = false
+    implicit val timeout = Timeout(10.seconds)
 
-  val system = ActorSystem("HearthStatsCompanion")
+    def encodeImages(images: List[TimedImage]): Future[Long] =
+      (actor ? images).mapTo[Long]
 
-  case class Encode(bi: BufferedImage)
-  case object Finish
+    /**
+     * Returns the name of the compressed file.
+     */
+    def finish(): Future[String] =
+      actor.ask(Finish)(60 seconds).asInstanceOf[Future[String]]
 
-  val encodeActor = system.actorOf(Props(new Actor {
-    def receive = {
-      case Finish => sender ! finishImpl()
-      case Encode(bi) => encodeImageImpl(bi)
-    }
-  }))
+    case object Finish
 
-  def store(bi: BufferedImage): Unit = {
+    class EncodeActor extends Actor {
+      val video = File.createTempFile("HSReplay", ".mp4").getAbsolutePath
+      var closed = false
+      var writer: IMediaWriter = _
 
-  }
+      def receive = {
+        case images: List[TimedImage] =>
+          if (!closed) {
+            for (i <- images) encode(i.bi, i.timeMs)
+          }
+          sender ! images.map(_.timeMs).max
+        case Finish =>
+          if (!closed) {
+            closed = true
+            writer.close()
+            info(s"Video $video finished encoding")
+            sender ! video
+          }
+      }
 
-  override def encodeImage(bi: BufferedImage): Unit =
-    if (!closed) {
-      receivedFrames += 1
-      debug(s"$receivedFrames received to encode")
-      encodeActor ! Encode(bi)
-    }
+      private def createWriter = {
+        val writer = ToolFactory.makeWriter(video)
+        writer.addVideoStream(0, 0, IRational.make((framesPerSec * 1000).toInt, 1000), videoWidth, videoHeight)
+        writer
+      }
 
-  def encodeImageImpl(bi: BufferedImage): Unit = {
-    val f = new File(picFolder.toFile, "t" + System.nanoTime)
-    val resized = resize(bi, videoWidth, videoHeight)
-    ImageIO.write(resized, "png", f)
-    storedFrames += 1
-    debug(s"$storedFrames stored")
-  }
+      var lastTime = 0
 
-  /**
-   * Returns the name of the compressed file.
-   */
-  override def finish(): Future[String] = if (!closed) {
-    closed = true
-    encodeActor.ask(Finish)(Timeout(10.seconds)).map(_.toString)
-  } else Promise[String].future // never completes
+      private def encode(bi: BufferedImage, timeMs: Long) = {
+        try {
+          val encodingStart = System.nanoTime
+          val image = new BufferedImage(bi.getWidth, bi.getHeight, BufferedImage.TYPE_3BYTE_BGR)
+          image.getGraphics().drawImage(bi, 0, 0, null)
+          if (writer == null) {
+            writer = createWriter
+            val w = image.getWidth
+            val h = image.getHeight
+            info(s"writing to $video : ${w}x$h @$framesPerSec")
+          }
+          lastTime += (1000 / framesPerSec).toInt
+          writer.encodeVideo(0, image, lastTime, TimeUnit.MILLISECONDS)
+          val duration = (System.nanoTime - encodingStart) / 1000000
+          debug(s"encoded until $lastTime ms, took $duration ms")
+        } catch {
+          case NonFatal(e) => warn(s"could not encode an image into video", e)
+          //normally only happens with screenshots used in tests
+        }
 
-  private def buildVideoFromImages(): Unit = {
-    for (f <- picFolder.toFile.listFiles.sortBy(_.getName)) {
-      val img = ImageIO.read(f)
-      val toEncode = Picture.create(img.getWidth, img.getHeight, ColorSpace.YUV420)
-      for (i <- 0 until 3) Arrays.fill(toEncode.getData()(i), 0)
-      transform.transform(AWTUtil.fromBufferedImage(img), toEncode)
-      _out.clear()
-      val result = encoder.encodeFrame(_out, toEncode)
-      spsList.clear()
-      ppsList.clear()
-      H264Utils.encodeMOVPacket(result, spsList, ppsList)
-      outTrack.addFrame(new MP4Packet(result, frameNo, framesPerSec.toLong, 1, frameNo, false, null, frameNo, 0))
-      frameNo += 1
-      debug(s"$frameNo encoded in video")
-    }
-  }
-
-  private def finishImpl(): String = {
-    buildVideoFromImages()
-    outTrack.addSampleEntry(H264Utils.createMOVSampleEntry(spsList, ppsList))
-    muxer.writeHeader()
-    NIOUtils.closeQuietly(ch)
-    val reader = ToolFactory.makeReader(out.getAbsolutePath)
-    val compressed = File.createTempFile("video", ".mp4").getAbsolutePath
-    reader.addListener(ToolFactory.makeWriter(compressed, reader))
-    while (reader.readPacket() == null) ()
-    out.delete()
-    compressed
-  }
-
-  private def resize(bi: BufferedImage, x: Int, y: Int): BufferedImage = {
-    val (h, w) = (bi.getHeight, bi.getWidth)
-    if (w <= x && h <= y) bi
-    else {
-      val scale = Math.min(x.toFloat / w, y.toFloat / h)
-      val at = AffineTransform.getScaleInstance(scale, scale)
-      val scaleOp = new AffineTransformOp(at, AffineTransformOp.TYPE_BILINEAR)
-      scaleOp.filter(bi, null)
+      }
     }
   }
 }
